+++ packages/api/test/__fixtures__/sdk/operationid-quirks/index.ts
@@ -1,6 +1,6 @@
 import Oas from 'oas';
 import APICore from 'api/dist/core';
-import definition from './openapi.json';
+import definition from '../../../__fixtures__/definitions/operationid-quirks.json';
 
 export default class SDK {
   spec: Oas;
