+++ packages/api/test/__fixtures__/sdk/petstore/index.ts
@@ -1,6 +1,6 @@
 import Oas from 'oas';
 import APICore from 'api/dist/core';
-import definition from './openapi.json';
+import definition from '@readme/oas-examples/3.0/json/petstore.json';
 
 export default class SDK {
   spec: Oas;
