+++ packages/api/test/__fixtures__/sdk/simple-js-cjs/index.js
@@ -6,11 +6,11 @@ var __importDefault =
   };
 var oas_1 = __importDefault(require('oas'));
 var core_1 = __importDefault(require('api/dist/core'));
-var openapi_json_1 = __importDefault(require('./openapi.json'));
+var simple_json_1 = __importDefault(require('../../../__fixtures__/definitions/simple.json'));
 var SDK = /** @class */ (function () {
   function SDK() {
     this.authKeys = [];
-    this.spec = oas_1.default.init(openapi_json_1.default);
+    this.spec = oas_1.default.init(simple_json_1.default);
     this.core = new core_1.default(this.spec, 'simple-js-cjs/1.0.0 (api/5.0-unit-testing)');
   }
   /**
