+++ packages/api/test/__fixtures__/sdk/simple-js-esm/index.js
@@ -1,6 +1,6 @@
 import Oas from 'oas';
 import APICore from 'api/dist/core';
-import definition from './openapi.json';
+import definition from '../../../__fixtures__/definitions/simple.json';
 export default class SDK {
   constructor() {
     this.authKeys = [];
