+++ packages/api/test/cli/codegen/languages/typescript.test.ts
@@ -1,4 +1,6 @@
 /* eslint-disable import/first */
+import type { TSGeneratorOptions } from '../../../../src/cli/codegen/languages/typescript';
+
 import chai, { expect } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
@@ -22,6 +24,16 @@ import TSGenerator from '../../../../src/cli/codegen/languages/typescript';
 
 chai.use(chaiPlugins);
 chai.use(sinonChai);
+
+function assertSDKFixture(file: string, fixture: string, opts: TSGeneratorOptions = {}) {
+  return async function () {
+    const oas = await import(file).then(Oas.init);
+    await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
+
+    const ts = new TSGenerator(oas, file, fixture, opts);
+    expect(await ts.generator()).toMatchSDKFixture(fixture);
+  };
+}
 
 describe('typescript', function () {
   describe('#installer', function () {
@@ -53,66 +65,38 @@ describe('typescript', function () {
   });
 
   describe('#generator', function () {
-    it('should generate typescript (by default)', async function () {
-      const oas = await import('../../../__fixtures__/definitions/simple.json').then(Oas.init);
-      await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
+    it(
+      'should generate typescript (by default)',
+      assertSDKFixture('../../../__fixtures__/definitions/simple.json', 'simple-ts')
+    );
 
-      const ts = new TSGenerator(oas, './openapi.json', 'simple-ts');
-      expect(await ts.generator()).toMatchSDKFixture('simple-ts');
-    });
+    it(
+      'should be able to generate valid TS when a body is optional but metadata isnt',
+      assertSDKFixture('../../../__fixtures__/definitions/optional-payload.json', 'optional-payload')
+    );
 
-    it('should be able to generate valid TS when a body is optional but metadata isnt', async function () {
-      const oas = await import('../../../__fixtures__/definitions/optional-payload.json').then(Oas.init);
-      await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
+    it('should work against the petstore', assertSDKFixture('@readme/oas-examples/3.0/json/petstore.json', 'petstore'));
 
-      const ts = new TSGenerator(oas, './openapi.json', 'optional-payload');
-      expect(await ts.generator()).toMatchSDKFixture('optional-payload');
-    });
+    it('should work against our OAS', assertSDKFixture('@readme/oas-examples/3.0/json/readme.json', 'readme'));
 
-    it('should work against the petstore', async function () {
-      const oas = await import('@readme/oas-examples/3.0/json/petstore.json').then(Oas.init);
-      await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
-
-      const ts = new TSGenerator(oas, './openapi.json', 'petstore');
-      expect(await ts.generator()).toMatchSDKFixture('petstore');
-    });
-
-    it('should work against our OAS', async function () {
-      const oas = await import('@readme/oas-examples/3.0/json/readme.json').then(Oas.init);
-      await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
-
-      const ts = new TSGenerator(oas, './openapi.json', 'readme');
-      expect(await ts.generator()).toMatchSDKFixture('readme');
-    });
-
-    it('should handle some quirky `operationId` cases', async function () {
-      const oas = await import('../../../__fixtures__/definitions/operationid-quirks.json').then(Oas.init);
-      await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
-
-      const ts = new TSGenerator(oas, './openapi.json', 'operationid-quirks');
-      expect(await ts.generator()).toMatchSDKFixture('operationid-quirks');
-    });
+    it(
+      'should handle some quirky `operationId` cases',
+      assertSDKFixture('../../../__fixtures__/definitions/operationid-quirks.json', 'operationid-quirks')
+    );
 
     describe('javascript generation', function () {
-      it('should generate a CommonJS library', async function () {
-        const oas = await import('../../../__fixtures__/definitions/simple.json').then(Oas.init);
-        await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
+      it(
+        'should generate a CommonJS library',
+        assertSDKFixture('../../../__fixtures__/definitions/simple.json', 'simple-js-cjs', { outputJS: true })
+      );
 
-        const ts = new TSGenerator(oas, './openapi.json', 'simple-js-cjs', { outputJS: true });
-        expect(await ts.generator()).toMatchSDKFixture('simple-js-cjs');
-      });
-
-      it('should generate am ESM library', async function () {
-        const oas = await import('../../../__fixtures__/definitions/simple.json').then(Oas.init);
-        await oas.dereference({ preserveRefAsJSONSchemaTitle: true });
-
-        const ts = new TSGenerator(oas, './openapi.json', 'simple-js-esm', {
+      it(
+        'should generate am ESM library',
+        assertSDKFixture('../../../__fixtures__/definitions/simple.json', 'simple-js-esm', {
           outputJS: true,
           compilerTarget: 'esm',
-        });
-
-        expect(await ts.generator()).toMatchSDKFixture('simple-js-esm');
-      });
+        })
+      );
     });
   });
 });
